refactor: Move the Poperty Copy function to a util
